[refactor] add template button
