+++ renderer/components/buttons/MinusBtn.tsx
@@ -1,7 +1,7 @@
-import { Sprite } from '@inlet/react-pixi'
 import { InteractionEvent } from 'pixi.js'
 
 import { BasicSpriteProps } from '../../types/sprite'
+import ButtonTemplate from './template/ButtonTemplate'
 
 interface Props extends BasicSpriteProps {
   handleClick: (event: InteractionEvent) => void
@@ -9,14 +9,12 @@ interface Props extends BasicSpriteProps {
 
 const MinusBtn = ({ x, y, scale = 1, handleClick }: Props) => {
   return (
-    <Sprite
+    <ButtonTemplate
       image="/static/img/ShopSelectNum/minus.png"
       x={x}
       y={y}
       scale={scale}
-      interactive={true}
-      buttonMode={true}
-      click={handleClick}
+      handleClick={handleClick}
     />
   )
 }
