+++ renderer/components/buttons/OptionBtn.tsx
@@ -1,33 +1,22 @@
-import { Sprite } from '@inlet/react-pixi'
 import { InteractionEvent } from 'pixi.js'
-import { useState } from 'react'
 
 import { BasicSpriteProps } from '../../types/sprite'
-import { containsPoint } from '../../utils/PixiAPI'
+import ButtonTemplate from './template/ButtonTemplate'
 
 interface Props extends BasicSpriteProps {
-  handleSettingClick: (event: InteractionEvent) => void
+  handleClick: (event: InteractionEvent) => void
 }
 
-const SettingBtn = ({ x = 0, y = 0, scale = 1, handleSettingClick }: Props) => {
-  const [alpha, setAlpha] = useState(1)
-  const mouseover = () => setAlpha(0.8)
-  const mouseout = () => setAlpha(1)
+const OptionBtn = ({ x = 0, y = 0, scale = 1, handleClick }: Props) => {
   return (
-    <Sprite
+    <ButtonTemplate
       image="/static/img/option-btn.png"
       x={x}
       y={y}
       scale={scale}
-      interactive={true}
-      click={handleSettingClick}
-      buttonMode={true}
-      mouseover={mouseover}
-      mouseout={mouseout}
-      alpha={alpha}
-      containsPoint={containsPoint}
+      handleClick={handleClick}
     />
   )
 }
 
-export default SettingBtn
+export default OptionBtn
