+++ renderer/components/buttons/QuitBtn.tsx
@@ -1,9 +1,7 @@
-import { Sprite } from '@inlet/react-pixi'
 import { InteractionEvent } from 'pixi.js'
-import { useState } from 'react'
 
 import { BasicSpriteProps } from '../../types/sprite'
-import { containsPoint, containsPointClickThrouth } from '../../utils/PixiAPI'
+import ButtonTemplate from './template/ButtonTemplate'
 
 interface Props extends BasicSpriteProps {
   isClickThrouth?: boolean
@@ -17,24 +15,14 @@ const QuitBtn = ({
   isClickThrouth = false,
   handleClick,
 }: Props) => {
-  const [alpha, setAlpha] = useState(1)
-  const mouseover = () => setAlpha(0.8)
-  const mouseout = () => setAlpha(1)
-
-  // 属性のみの切り替え方法が不明なので要素ごとわける
   return (
-    <Sprite
+    <ButtonTemplate
       image="/static/img/quit-btn.png"
       x={x}
       y={y}
       scale={scale}
-      interactive={true}
-      alpha={alpha}
-      click={handleClick}
-      mouseover={mouseover}
-      mouseout={mouseout}
-      containsPoint={isClickThrouth ? containsPointClickThrouth : containsPoint}
-      buttonMode={true}
+      handleClick={handleClick}
+      isClickThrouth={isClickThrouth}
     />
   )
 }
