+++ renderer/components/buttons/ShopBtn.tsx
@@ -1,9 +1,7 @@
-import { Sprite } from '@inlet/react-pixi'
 import { InteractionEvent } from 'pixi.js'
-import { useState } from 'react'
 
 import { BasicSpriteProps } from '../../types/sprite'
-import { containsPoint, containsPointClickThrouth } from '../../utils/PixiAPI'
+import ButtonTemplate from './template/ButtonTemplate'
 
 interface Props extends BasicSpriteProps {
   isClickThrouth?: boolean
@@ -17,23 +15,14 @@ const ShopBtn = ({
   isClickThrouth = false,
   handleClick,
 }: Props) => {
-  const [alpha, setAlpha] = useState(1)
-  const mouseover = () => setAlpha(0.8)
-  const mouseout = () => setAlpha(1)
-
   return (
-    <Sprite
+    <ButtonTemplate
       image="/static/img/shop-btn.png"
       x={x}
       y={y}
       scale={scale}
-      interactive={true}
-      alpha={alpha}
-      click={handleClick}
-      mouseover={mouseover}
-      mouseout={mouseout}
-      containsPoint={isClickThrouth ? containsPointClickThrouth : containsPoint}
-      buttonMode={true}
+      handleClick={handleClick}
+      isClickThrouth={isClickThrouth}
     />
   )
 }
