+++ renderer/pages/feed.tsx
@@ -116,7 +116,7 @@ const Feed = () => {
         y={950}
         scale={0.5}
         isZero={isEmpty}
-        handleClickFish={handleClickFish}
+        handleClick={handleClickFish}
       />
       <CuteFish x={1795} y={928} scale={0.2} />
       <Text
